# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TontiTOGO API",
        "description": "API RESTful de la plateforme TontiTOGO — Gestion numérique des tontines au Togo avec suivi des cotisations et notifications SMS.",
        "contact": {
            "name": "Support TontiTOGO",
            "email": "support@tontitogo.tg"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur local de développement"
        }
    ],
    "paths": {
...
        "/api/responsible/register": {
            "post": {
                "tags": [
                    "Auth Responsable"
                ],
                "summary": "Inscription d'un nouveau responsable",
                "operationId": "4f2b832b808b2fc3508dc71aba4d7892",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "organization_name",
                                    "firstname",
                                    "lastname",
                                    "phone",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "organization_name": {
                                        "type": "string",
                                        "example": "Tontine Solidarité"
                                    },
                                    "firstname": {
                                        "type": "string",
                                        "example": "Kofi"
                                    },
                                    "lastname": {
                                        "type": "string",
                                        "example": "Mensah"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "22890123456"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "kofi@example.tg"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compte créé avec succès"
                    },
                    "422": {
                        "description": "Erreurs de validation"
                    }
                }
            }
        },
        "/api/responsible/login": {
            "post": {
                "tags": [
                    "Auth Responsable"
                ],
                "summary": "Connexion du responsable",
                "operationId": "77a5c2dba899f6934e9ba31950dbc5ea",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login",
                                    "password"
                                ],
                                "properties": {
                                    "login": {
                                        "description": "Email ou téléphone",
                                        "type": "string",
                                        "example": "kofi@example.tg"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie — retourne user + token"
                    },
                    "422": {
                        "description": "Identifiants incorrects"
                    },
                    "403": {
                        "description": "Compte suspendu"
                    }
                }
            }
        },
        "/api/responsible/logout": {
            "post": {
                "tags": [
                    "Auth Responsable"
                ],
                "summary": "Déconnexion du responsable",
                "operationId": "8f977318768239eb6008a9fe84b6105b",
                "responses": {
                    "200": {
                        "description": "Déconnecté"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/responsible/me": {
            "get": {
                "tags": [
                    "Auth Responsable"
                ],
                "summary": "Profil du responsable authentifié",
                "operationId": "eac97702e5e5c22b1241fdfeaedbe24f",
                "responses": {
                    "200": {
                        "description": "Profil retourné"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/responsible/profile": {
            "put": {
                "tags": [
                    "Auth Responsable"
                ],
                "summary": "Mettre à jour le profil du responsable",
                "operationId": "b6a0005d4a87acb5f073fb68ced1a508",
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "firstname": {
                                        "type": "string"
                                    },
                                    "lastname": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "22890123456"
                                    },
                                    "current_password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "avatar": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil mis à jour"
                    },
                    "422": {
                        "description": "Erreurs de validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/responsible/account": {
            "delete": {
                "tags": [
                    "Auth Responsable"
                ],
                "summary": "Supprimer définitivement son compte (droit à l'oubli)",
                "operationId": "c046f357d43bbf788466b89a997b0c5b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "password"
                                ],
                                "properties": {
                                    "password": {
                                        "description": "Mot de passe actuel pour confirmation",
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Compte supprimé définitivement"
                    },
                    "422": {
                        "description": "Mot de passe incorrect"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Entrez votre token Sanctum : Bearer {token}"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}